+++ lib/barista.js
@@ -3,8 +3,8 @@
 const Rule = require('./rule');
 const Output = require('./output');
 const sass = require('./utils/sass');
-const defaultOptions = require('./defaults/options');
-const defaultOutputOptions = require('./defaults/outputOptions');
+const defaultOptions = require('./defaults/barista');
+const defaultOutputOptions = require('./defaults/output');
 const cssom = require('./utils/cssom');
 const resolve = require('./utils/resolve');
 const isValid = require('./utils/isValid');
