+++ lib/defaults/output.js
@@ -1,4 +1,4 @@
-const defaultOptions = require('./options');
+const defaultOptions = require('./barista');
 
 const propTypeFn = () => {
   return false;
