+++ lib/output.js
@@ -1,10 +1,11 @@
 // Output
 'use strict';
 
-var renderer = require('./renderer');
 var genki = require('genki');
+var renderer = require('./utils/renderer');
+var defaultOptions = require('./defaults/output');
 
-var BaristaOutput = function(options) {
+var BaristaOutput = function(options = defaultOptions) {
   this.$ = false;
   this.css = false;
   this.data = false;
