+++ lib/utils/CSSOM.js
@@ -1,7 +1,7 @@
 const postcss = require('postcss');
 
 const parse = (css = false) => {
-  if (!css) {
+  if (!css || typeof css !== 'string') {
     return false;
   }
   return postcss.parse(css);
@@ -9,5 +9,4 @@ const parse = (css = false) => {
 
 module.exports = {
   parse,
-  postcss,
 };
