+++ lib/utils/renderer.js
@@ -1,9 +1,6 @@
-// Renderer
-'use strict';
+const emmet = require('../vendor/emmet');
 
-var emmet = require('../vendor/emmet');
-
-var sanitize = function(selectors) {
+const sanitize = function(selectors) {
   return selectors.trim()
     .replace(/\ > \ /g, ' ')
     .replace(/>/g, ' ')
@@ -13,7 +10,7 @@ var sanitize = function(selectors) {
     .replace(/\ /g, ' > ');
 };
 
-var renderer = function(document, selectors) {
+const renderer = function(document, selectors) {
   if (!document) {
     return;
   }
