+++ lib/utils/resolve.js
@@ -1,19 +1,19 @@
 const harvester = require('seed-harvester');
 const pathfinder = require('sass-pathfinder');
 const path = require('path');
-const options = require('../defaults/options');
+const defaultOptions = require('../defaults/barista');
 
 const stripRoot = (p = '') => {
   const {
     root,
-  } = options;
+  } = defaultOptions;
   return p.replace(new RegExp(root, 'g'), '');
 }
 
 const pathname = (p = '') => {
   const {
     root,
-  } = options;
+  } = defaultOptions;
   return path.join(root, stripRoot(p));
 };
 
@@ -36,7 +36,7 @@ const seedPaths = (pack = false) => {
   }
 };
 
-const sassPaths = (o = options) => {
+const sassPaths = (o = defaultOptions) => {
   const paths = [includePaths(o.includePaths)];
   if (o.includeSeedPaths) {
     paths.push(seedPaths(o.pack));
