+++ lib/utils/sass.js
@@ -2,7 +2,7 @@ const fs = require('fs');
 const nodeSass = require('node-sass');
 const path = require('path');
 const pathfinder = require('sass-pathfinder');
-const defaultOptions = require('../defaults/options');
+const defaultOptions = require('../defaults/barista');
 
 const sassOptions = (o = defaultOptions) => {
   const {
